Removed a lot of warnings.

+++ AGEproject/src/Game.cpp
@@ -16,7 +16,7 @@ void Game::Initialise()
 	free_cam = new Entity;
 	auto cam = std::make_unique<Free_Camera>(glm::half_pi<float>());
 	cam->SetPosition(glm::dvec3(10.0, 5.0, 50.0));
-	cam->SetProjection(GameEngine::Get().GetScreenWidth() / GameEngine::Get().GetScreenHeight(), 2.414f, 1000);
+	cam->SetProjection((float)(GameEngine::Get().GetScreenWidth() / GameEngine::Get().GetScreenHeight()), 2.414f, 1000);
 	free_cam->AddComponent(move(cam));
 
 	// Add a red point light to 0, 0.5, 0
@@ -141,7 +141,7 @@ void Game::Update()
 void Game::Render()
 {
 	// Clear the opengl buffer.
-	glClearColor(0.1, 0.0, 0.4, 1);
+	glClearColor(0.1f, 0.0f, 0.4f, 1.0f);
 	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 	//glDisable(GL_CULL_FACE);
